+++ NEWS.md
@@ -1,3 +1,10 @@
+# astrocalculation 0.1.5 -- 2025-12-24
+* Expand R6 data manager to handle user profile management
+* First logic saving user profile data
+
+# astrocalculation 0.1.4 -- 2025-12-21
+* initialize postgres db (local db for development)
+
 # astrocalculation 0.1.3 -- 2025-12-13
 * export add_datetime and minus_datetime